fix: dev: make tests about time more stable
This test was asserting that two consecutives calls of the `ts` property return
different result, the first call being lower than the second call.

But the value returned by `ts` is a floating point number expressed in seconds,
and it seems that, sometimes, Python's so fast that the assertion fails.
Using the `test` command with `--repeat 1000`, it seems that I have 2% runs
which are failing with this test on my computer.

The fix should be sufficient, since each doctest lines are reinterpreted by
Doctest, and I guess it should have a sufficient cost so that it now passes
at 100%.